requirements.txt
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
plotly==5.15.0
Requests==2.31.0
scikit_learn==1.2.2
streamlit==1.26.0
streamlit_lottie==0.0.5
